LPS-72820 Moves tests
