+++ modules/apps/web-experience/layout/layout-test/src/testIntegration/java/com/liferay/layout/service/test/LayoutFriendlyURLServiceTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.service;
+package com.liferay.layout.service.test;
 
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.Layout;
