+++ modules/apps/web-experience/layout/layout-test/src/testIntegration/java/com/liferay/layout/verify/test/VerifyLayoutTest.java
@@ -12,10 +12,12 @@
  * details.
  */
 
-package com.liferay.portal.verify;
+package com.liferay.layout.verify.test;
 
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
+import com.liferay.portal.verify.VerifyLayout;
+import com.liferay.portal.verify.VerifyProcess;
 import com.liferay.portal.verify.test.BaseVerifyProcessTestCase;
 
 import org.junit.ClassRule;
